Fix inconsistently renamed guard tabbards/cloaks. (#1965)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
This PR does as it says, it fixes how inconsistent guard tabbards/cloaks
are named. Sometimes you'll get it named, other times you don't. Now if
you latejoin/roundstart or whatever it always names it at least from
what i've tested.

This is for the tabbards/cloaks on the roles- Knights, City Guards, Men
at arms, and Gatemasters.

## Why It's Good For The Game
Now that helmets cover guards face again, we will want the
tabbard/cloaks being named consistantly. Unknown guards suck, especially
more so with guards that work in malice.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->