(SOUND) Fixes Astratan Templar's Music (#2057)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

Gives the CORRECT combat theme to the Astratan Templar because they were
lacking it. I swear to god I did this before, I swear to god. This is
like the second time. I'm losing it.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

It was always supposed to be like this.

Also I fucked up the formatting on this I guess because it's the
asscrack of midnight but please stop making me fix the Astratan song, I
don't know what I did to spite god.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->