Influence adjustment, slurs and h*nd h*lding (#2062)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Replace astratas tracked influence taxes collected with slurs spoken,
slurs spoken is checked when a mob with a client speaks compared against
a new list.

If the mob says a slur NOT for their own species they are tracked as a
speciesist for round end stats as well.

Replace eora's clingy people stat with "H\*nds H\*ld".

Shout out to OOC and D&D for help with the slur list. Comment new slurs
down below.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->