Replaces Rakshari Thieves' Cant with Zybean (#2095)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Replaces the Rakshari knowledge of Thieves' Cant with Zybean instead.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

Rakshari are desert cats from Zybantine with centuries of close
integration into its culture, as stated in their description. They are
eligible for most Zybantine roles. They have enormous roleplaying
potential for interacting with Zybantine roles as foreigners. And yet
tragically, they can only speak a thieves' sign language only understood
by them.

Thieves' Cant is a language basically without any use for Rakshari due
to the lack of a thieves' guild, but more importantly it is also just
not very useable for roleplay. Rakshari are meant to be discriminated
against as slavers and outsiders but don't feel especially "foreign".
With their Zybantine descent, Rakshari do not really need a unique
language of their own - being able to speak Zybean would fit their lore
just fine and allow more opportunities for Zybantine character roleplay
as well as giving the locals another reason to be suspicious of them.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->