(SOUND) Gives the Inquisitor's Duster unique footsteps (#2087)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds unique footsteps for the Inquisitor's Duster.

Foley sourced from:
[https://freesound.org/people/jarhead123/packs/31306/](https://freesound.org/people/jarhead123/packs/31306/)

## Why It's Good For The Game

The Inquisitor is meant to be a menacing figure all-around. It should be
very obvious when he's approaching - anyone sane should be shitting
bricks when they hear him rounding a corner, or if he's on a different
Z-Level entirely.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->