Pilgrim Noble Race-lock change (#2108)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Allows any races except hollow-kin/tieflings to be a foreign noble.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Tieflings is due to lore reason, Hollow-kin because they're pets. and it
make sense, A foreigner noble could legit be any races, are you telling
me we can't have a rakshari noble that came from zybantine? real
strange.

![image](https://github.com/user-attachments/assets/b7616dd9-020b-4d39-a948-b5d47754751e)

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->