Elven War Club (#2111)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Adds two new weapons, courtesy of 7erracotta.

![image](https://github.com/user-attachments/assets/4c5ad614-ba59-4508-ae59-2b0e942046d9)

A gunstock war club but for elves, Similar to the mace but has reduced
damages for the iron version but lowered str requirement too, can't
smash but can chop and cut, sharp too
added recipes to craft it
Clerics elves get iron version.
Black oak mercs can now choose between the steel version of the mace or
their regular spear.


![image](https://github.com/user-attachments/assets/d46f32f4-f9da-4c08-aaa4-3cf5e27b9596)

![image](https://github.com/user-attachments/assets/34bd9221-be3a-48c9-a8c0-01798afed32b)


![image](https://github.com/user-attachments/assets/78b64ca1-701e-4215-b939-28fcffde1027)

![image](https://github.com/user-attachments/assets/a72b5d72-84d6-4b44-a741-43545398d375)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
oh y'know, it'll be nice to have new weapon stuffs.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->