Wooden stool buff (#2126)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Makes wooden stool crafting difficulty 0.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
The humble wooden stool gets the short end of the stick in the chair
tech tree. Railings, ladders, walls and tables are all fair game for a
level 0 carpenter, but stools are not. What use is a table without a
seat? Other sittable(?) furniture such as the wooden chair and fancy
wooden chair require increasingly higher carpentry skill to craft - but
the lowly stool is alone in sharing the same skill level requirement as
its upgraded form, the chair.

Having a simple place to sit down that doesn't first require journeyman
carpentry training would mildly improve comfort and reduce lower back
pain for everybody.

Positives include:
- Fishers can sit down and relax at their favourite fishing spot!
- Beggars can have somewhere to rest by the campfire!
- Build a couple of seats and a table and play some card games wherever
you like!
- Slightly less awkward than having to eat at the table standing up!
- Non-carpenters using a noose get the extra luxury of a stool to kick
over!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->