Vanderlin and Daftmarsh map tweaks. (#2140)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
-Restores keep passages to their former glory before map updates,
expands in some areas.
-Adds additional cutlery and bowls to servants' quarters.
-Daftmarsh bandit signs.
-Daftmarsh secure door for often-raided merchant boat.
-Removes Daftmarsh gators, as they cause insane TiDi.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
A lot of the pasage utility was lost during map updates and iterations.
Fixes TiDi issues on Daftmarsh until gators can be fixed. Adds some
missing posters.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>