Fixes foreigners not getting their recruited title (#1983)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

I don't know how long it has been like this but quite a while already,
this fixes the issues of people recruiting foreigners and they lacking
any title when you exam, the truth is that their title did indeed
change, it was just not displaying because of the foreigner trait
itself.
This PR creates a new trait called "Recruited", all foreigners who get
recruited will gain this new trait and will have their new title
displayed without any issues:

![image](https://github.com/user-attachments/assets/14f85d3c-2cbd-46a7-a6e2-40aeeb81727f)

Closes:#1360

This include all other recruitments for foreigners.
## Why It's Good For The Game

Fix!

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->