(SOUND) New Bandit Mage theme. (#2132)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

[Adds the song "(APAB) All Paladins Are Bastards" by Burialgoods. Thank
you for your dutiful service to
knavemaxxing.](https://www.youtube.com/watch?v=3Z_cOxF1o48)

Adds it to Bandit Mage, replacing the previous song and deprecating it.

![he fucking did
it](https://github.com/user-attachments/assets/3703d3e3-e587-47cb-b6ed-4c4040a203b2)


## Why It's Good For The Game

Fuck the King's demesne, I'm not a big fan of the monarchy.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->