Adds the new Gatemaster Outfit. (#2114)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Adds the new gatemaster outfit that I've had done for a while now, just
coded it into the game.
This also makes the OLD gatemaster helmet be used for the town watch
archers as I intended.

screenshot (forgot to get the complete screenshot in game while testing,
this discord one will do)

![image](https://github.com/user-attachments/assets/5dfec2d1-3a7a-4827-a917-7651cf3edaa2)

## Why It's Good For The Game
It looks cool.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->