Rosewood Jousting and Inn Update (#2117)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request


![image](https://github.com/user-attachments/assets/7d9b8500-4770-420d-ac1a-c1c05eb80ad1)

![image](https://github.com/user-attachments/assets/b63ba074-9d3b-471b-8e22-6fb27a0bdb77)

![image](https://github.com/user-attachments/assets/84d1efcf-f2c5-40e2-bb1d-2cfedbf40f89)

Adds in the new jousting arena and inn rework as desired by Drag. The
bar now hosts and island design that allows the innkeep to serve and
keep an eye on all in the tavern. The kitchen now is directly behind the
bar with a staircase that goes down into a basement similar to
vanderlins. Wine is held in a more secure storage as well.

Jousting arena was added with a new way to fight and an incentive for
butchers to breed saigas for jousting! Be a knight and take the bronze
lances in the jousting arena's basement to fight one another. Careful
though the points are deadly!

Fixed a few roads to make them less snowy
Fixed warehouse issue with lighthouse
Fixed double SCOMs in orphanage
Fixed lighting issue in orphanage

## Why It's Good For The Game

Overall change to rosewood's inn with the hopes to incentivize more
people to fight one another in a knightly order rather than just
brawling it out in the pit. This should also hopefully make nobles duel
or have knights duel one another on saigas.
## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->